casting per evitare errori

+++ probing/LinearProber.py
@@ -10,14 +10,14 @@ class LogisticRegression(nn.Module):
         self.linear = nn.Linear(input_dim, 1, bias=use_bias)
 
     def forward(self, x):
-        return torch.sigmoid(self.linear(x)).squeeze()
+        return torch.sigmoid(self.linear(x.float())).squeeze()
 
     @torch.inference_mode()
     def predict(self, x):
         if not isinstance(x, torch.Tensor):
             x = torch.tensor(x)
         x = x.to(self.linear.weight.device)
-        return torch.sigmoid(self.linear(x)).squeeze()
+        return torch.sigmoid(self.linear(x.float())).squeeze()
 
 
 class LinearProber:
